Update CombinePath function to include pathSeparator parameter
The CombinePath function in the IOUtil.cs file has been updated to accept a pathSeparator parameter. This change allows for greater flexibility when defining the separator to use when combining paths. The corresponding uses of the function in MinecraftProcessBuilder.cs and IOUtilTests.cs have also been modified to include this new parameter.